feat: Add AutoMapper and configure mapping for CreateTodoRequest and UpdateTodoRequest
This commit adds the AutoMapper package reference and configures the mapping between the CreateTodoRequest and UpdateTodoRequest contracts to the Todo entity. It also ignores the Id property of the Todo entity during mapping.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.