feat: Add ITodoServices interface
This commit adds the ITodoServices interface, which defines the contract for interacting with the Todo services in the ShareBill application. It includes methods for getting all todos, getting a todo by id, creating a new todo, updating a todo, and deleting a todo.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.