feat: Add GetAllAsync method to TodoController
This commit adds the GetAllAsync method to the TodoController class. The method handles a GET request to retrieve all todo items from the database. It calls the _todoServices.GetAllAsync method and returns the retrieved todos in the response. If an error occurs, it returns a 500 status code with an error message.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.