refactor: Update TodoDbContext to use Npgsql.EntityFrameworkCore.PostgreSQL
This commit updates the TodoDbContext class to use Npgsql.EntityFrameworkCore.PostgreSQL instead of Microsoft.EntityFrameworkCore.SqlServer. The OnConfiguring method is modified to use the UseNpgsql method with the updated connection string from the DbSettings class.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.