refactor: Update Todo page component to use @page directive and add PageTitle component
This commit updates the Todo page component to use the @page directive with the "/todo" route. It also adds the PageTitle component to set the page title to "Todo". This change improves the routing and enhances the user experience.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.