feat: Add Todo page link to navigation menu
This commit adds a new link to the navigation menu for the Todo page. The link uses the NavLink component and points to the "/todo" route. This change improves the user experience by providing easy access to the Todo page.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.